# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ساعدني - API specification for the Oman help marketplace
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user (returns OTP, does NOT log in)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered, OTP generated for verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Request OTP login (phone only)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OTP generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResponse"
        "404":
          description: Phone not found
  /auth/admin-login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login with phone + PIN (no OTP)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "403":
          description: Invalid PIN
        "404":
          description: Phone not found
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and complete login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired OTP
        "403":
          description: Account disabled
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, phone, userType, createdAt, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        userType:
          type: string
          enum: [customer, helper, admin]
        area:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        ratingCount:
          type: integer
          minimum: 0
          description: Count calculated from helper_ratings; not a client-maintained counter.
        isActive:
          type: boolean
        isVerified:
          type: boolean
        isBlocked:
          type: boolean
        lastLogin:
          type: ["string", "null"]
        otpCode:
          type: ["string", "null"]
        otpCreatedAt:
          type: ["string", "null"]
        helperWelcomeMessageSentAt:
          type: ["string", "null"]
        createdAt:
          type: string
        helperInterests:
          type: ["string", "null"]
        preferredAreas:
          type: ["string", "null"]
        serviceAreas:
          type: array
          items:
            type: string
    RegisterInput:
      type: object
      required: [name, phone, userType]
      properties:
        name:
          type: string
        phone:
          type: string
        userType:
          type: string
          enum: [customer, helper]
        area:
          type: ["string", "null"]
        preferredAreas:
          type: array
          items:
            type: string
    LoginInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    VerifyOtpInput:
      type: object
      required: [phone, otp]
      properties:
        phone:
          type: string
        otp:
          type: string
    AdminLoginInput:
      type: object
      required: [phone, pin]
      properties:
        phone:
          type: string
        pin:
          type: string
    OtpRequestResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        otp:
          type: string
        isVerified:
          type: boolean
        isAdmin:
          type: boolean
        otpDelivery:
          type: string
          enum: [whatsapp]
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
...